+++ .github/workflows/build.yaml
@@ -2,8 +2,8 @@ name: Build and Publish Docker Image
 
 on:
   push:
-    branches:
-      - main
+    # branches:
+    #   - main
   pull_request:
 
 jobs:
@@ -15,6 +15,13 @@ jobs:
       id-token: write
 
     runs-on: ubuntu-latest
+
+    # Define the services that should be built.
+    strategy:
+      matrix:
+        service:
+          - datum-authorization-webhook
+
     steps:
     - name: Checkout repository
       uses: actions/checkout@v3
@@ -30,7 +37,7 @@ jobs:
       id: meta
       uses: docker/metadata-action@v5.5.1
       with:
-        images: ghcr.io/datum-cloud/datum
+        images: ghcr.io/datum-cloud/${{ matrix.service }}
         tags: |
           type=schedule
           type=ref,event=branch
@@ -40,11 +47,19 @@ jobs:
           type=semver,pattern={{major}}
           type=sha
 
-    - name: Build and push Docker image
+    - name: Build ${{ matrix.service }}
       id: push
       uses: docker/build-push-action@v6.7.0
       with:
         context: .
+        file: cmd/${{ matrix.service }}/Dockerfile
         push: true
         tags: ${{ steps.meta.outputs.tags }}
         labels: ${{ steps.meta.outputs.labels }}
+
+    - name: Generate artifact attestation
+      uses: actions/attest-build-provenance@v1
+      with:
+        subject-name: ghcr.io/datum-cloud/${{ matrix.service }}
+        subject-digest: ${{ steps.push.outputs.digest }}
+        push-to-registry: true
